Add Orion source code
Orion is a fork of Anti-Grain Geometry (AGG).  AGG is a high-quality
2D rendering library that is available under a permissive license
with no external dependencies.  AGG is no longer maintained so we
have created our own fork that we can modify as necessary.

We renamed the fork to avoid issues when linking with other libraries
that use their own copies of AGG (e.g., pdfium).